backend/requirements.txt
Flask
streamlit
PyPDF2
langchain
google-generativeai
faiss-cpu
python-dotenv
werkzeug
langchain-google-genai
